metro-back/requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
attrs==23.2.0
contourpy==1.2.1
cycler==0.12.1
Django==5.0.6
django-compat==1.0.15
django-excel-response2==3.0.6
django-rest-enumfield==0.2.0
django-six==1.0.5
djangorestframework==3.15.1
drf-spectacular==0.27.2
excel-base==1.0.4
fonttools==4.53.0
inflection==0.5.1
install==1.3.5
isoweek==1.3.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
matplotlib==3.9.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
screen==1.0.1
six==1.16.0
sqlparse==0.5.0
TimeConvert==3.0.13
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
xlwt==1.3.0